requirements.txt
django>=5.0.2
djangorestframework>=3.14.0
django-vite>=3.0.3
django-cors-headers>=4.3.1
django-rest-knox>=4.2.0
mod-wsgi>=5.0.0
psycopg>=3.1.12
whitenoise[brotli]>=6.6.0
python-dateutil>=2.9.0
requests>=2.32.3
# Generate API documentation
drf-spectacular>=0.27.2
# Image verification
pillow>=10.4.0
